+++ packages/components/src/components/input/input.types.ts
@@ -1,0 +1,7 @@
+export interface InputInputDetail {
+  value: string;
+}
+
+export interface InputChangeDetail {
+  value: string;
+}