+++ packages/components/src/index.ts
@@ -2,3 +2,4 @@ export * from './components/button/index.js';
 export * from './components/checkbox/index.js';
 export * from './components/radio/index.js';
 export * from './components/toggle/index.js';
+export * from './components/input/index.js';